+++ src/thread_pool_async.ml
@@ -11,45 +11,80 @@ exception Invalid_thread_count of int
 
 exception Pool_already_destroyed
 
+exception State_manipulation_error of exn
+
 type 'state t = {
-  pool : 'state worker Pipe.Reader.t * 'state worker Pipe.Writer.t;
+  reader : 'state worker Pipe.Reader.t;
+  writer : 'state worker Pipe.Writer.t;
   create : unit -> 'state;
   destroy : 'state -> unit;
-  threads : int;
+  thread_count : int;
 }
 
-let init ~name ~threads ~create ~destroy =
-  match threads with
-  | _ when threads <= 0 -> raise @@ Invalid_thread_count threads
-  | _ ->
-      let create_thread ~name ~create =
-        let%bind thread = In_thread.Helper_thread.create ~name () in
-        let%bind state = In_thread.run ~thread create in
-        return {thread; state}
-      in
-      (* No strict ordering on the heap. All items are equal *)
-      let reader, writer = Pipe.create () in
-      let elements = List.init threads ~f:ignore in
-      let add_to_pool () =
-        create_thread ~name ~create >>| Pipe.write_without_pushback writer
-      in
-      let%bind () = Deferred.List.iter elements ~how:`Parallel ~f:add_to_pool in
-      return {pool = reader, writer; create; destroy; threads}
+let try_in_thread ~thread f = try_with (fun () -> In_thread.run ~thread f)
+
+let destroy_pool {reader; writer; thread_count; destroy; _} =
+  let%bind states_destroyed =
+    List.init thread_count ~f:ignore
+    |> Deferred.List.map ~how:`Parallel ~f:(fun () ->
+           match%bind Pipe.read reader with
+           | `Eof -> return @@ Ok ()
+           | `Ok {thread; state} -> try_in_thread ~thread (fun () -> destroy state))
+    |> Deferred.map ~f:Result.all_unit
+  in
+  Pipe.close writer;
+  match states_destroyed with
+  | Ok () -> return ()
+  | Error exn -> raise @@ State_manipulation_error exn
+
+let recreate_state ({create; destroy; _} as pool) {thread; state} =
+  let destroy_pool_and_raise exn =
+    let%bind () = destroy_pool pool in
+    raise @@ State_manipulation_error exn
+  in
+  match%bind try_in_thread ~thread (fun () -> destroy state) with
+  | Ok () -> (
+      match%bind try_in_thread ~thread create with
+      | Ok state -> return {thread; state}
+      | Error exn -> destroy_pool_and_raise exn)
+  | Error exn -> destroy_pool_and_raise exn
+
+let init ~name ~threads:thread_count ~create ~destroy =
+  if thread_count <= 0
+  then raise @@ Invalid_thread_count thread_count
+  else
+    (* No strict ordering on the heap. All items are equal *)
+    let reader, writer = Pipe.create () in
+    let pool = {reader; writer; create; destroy; thread_count} in
+    let%bind threads =
+      List.init thread_count ~f:ignore
+      |> Deferred.List.map ~how:`Parallel ~f:(In_thread.Helper_thread.create ~name)
+    in
+    let create_state thread = try_in_thread ~thread create in
+    match%bind
+      threads
+      |> Deferred.List.map ~how:`Parallel ~f:create_state
+      |> Deferred.map ~f:Result.all
+    with
+    | Ok states ->
+        let add_worker (thread, state) =
+          Pipe.write_without_pushback writer {thread; state}
+        in
+        List.zip_exn threads states |> List.iter ~f:add_worker;
+        return pool
+    | Error exn ->
+        let%bind () = destroy_pool pool in
+        raise @@ State_manipulation_error exn
 
 let execute {thread; state} work = In_thread.run ~thread (fun () -> work state)
 
-let with_worker {pool = reader, writer; destroy; create; _} ?(retries = 0) job =
-  let recreate {thread; state} =
-    let%bind () = In_thread.run ~thread (fun () -> destroy state) in
-    let%bind state = In_thread.run ~thread create in
-    return @@ {thread; state}
-  in
+let with_worker ({reader; writer; _} as pool) ?(retries = 0) job =
   let run_once worker =
     let%bind result = try_with (fun () -> job worker) in
     let%bind worker =
       match result with
       | Ok (Ok _) -> return worker
-      | Ok (Error _) | Error _ -> recreate worker
+      | Ok (Error _) | Error _ -> recreate_state pool worker
     in
     return (result, worker)
   in
@@ -76,12 +111,4 @@ let with' pool ?(retries = 0) work =
   let job worker = execute worker work in
   with_worker pool ~retries job
 
-let destroy {pool = reader, writer; threads; destroy; _} =
-  let elements = List.init threads ~f:ignore in
-  let%bind () =
-    Deferred.List.iter elements ~how:`Parallel ~f:(fun () ->
-        match%bind Pipe.read reader with
-        | `Eof -> return ()
-        | `Ok {thread; state} -> In_thread.run ~thread (fun () -> destroy state))
-  in
-  return @@ Pipe.close writer
+let destroy = destroy_pool